♻️ move local command history types to store boundary
